requirements.txt
streamlit
pandas
elevenlabs
python-dotenv
